+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### Added
-- More updates to languages
+- Sifting no longer has a small chance of returning sand
 
 ### Fixed
 - Missing festive variant for megaloceros
@@ -8,3 +8,4 @@
 - Config option generateVolcanoBiome not working on forge
 
 ### Mod Compatibility
+- Sifter and Analyzer recipes can now return nothing with a custom probability